feat(mir): close semantic input preconditions

+++ crates/vela_bytecode/src/compiler/hir_lowering/call_placements.rs
@@ -397,7 +397,7 @@ impl Compiler<'_, '_> {
                     TypeContractContext::NativeParameter {
                         function: method.to_owned(),
                         name: mutation_arg_debug_name(method, &neutral_param.name, index),
-                        index: u16::try_from(index).unwrap_or(u16::MAX),
+                        index: u32::try_from(index).unwrap_or(u32::MAX),
                     },
                     callback_shapes.as_deref().unwrap_or(&[]),
                 )
