+++ crates/vela_bytecode/src/compiler/hir_lowering/calls.rs
@@ -811,7 +811,7 @@ impl Compiler<'_, '_> {
             TypeContractContext::NativeParameter {
                 function: function.to_owned(),
                 name: param.name.clone(),
-                index: u16::try_from(index).unwrap_or(u16::MAX),
+                index: u32::try_from(index).unwrap_or(u32::MAX),
             }
         };
         self.compile_hir_expression_for_expected_type(
